requirements.txt
discord.py
asyncio

jamo